% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMCMC.R
\name{calculateApproxLikelihood}
\alias{calculateApproxLikelihood}
\title{Function to calculate the approximate likelihood based on weights}
\usage{
calculateApproxLikelihood(weights)
}
\arguments{
\item{weights}{vector with weights}
}
\value{
approximate likelihood
}
\description{
Function to calculate the approximate likelihood based on weights
}

