% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_PMCMC.R
\name{DE_PMCMC}
\alias{DE_PMCMC}
\title{Differential-Evolution MCMC zs with a particle filter to approximate the likelihood}
\usage{
DE_PMCMC(iterations, model, likelihood, prior, f0, startValues = NULL,
  numParticles, data, names = NULL, reportIntervall = 10,
  referenceData = NULL, parallel = FALSE, parallelOptions = list(packages
  = NULL, objects = NULL), pSnooker = 0.1, burnin = 0, thin = 1,
  f = 2.38, eps = 0, pGamma1 = 0.1, eps.mult = 0.2, eps.add = 0,
  zUpdateFrequency = 10)
}
\arguments{
\item{iterations}{iterations to run}

\item{model}{model to make predictions about the state}

\item{likelihood}{likelihood function (see details)}

\item{prior}{a prior created by \code{\link{BayesianTools::createPrior}}}

\item{numParticles}{number of particles in the particle filter}

\item{data}{vector with data}

\item{reportIntervall}{intervall in which the sampler's progress is printed}

\item{referenceData}{matrix with reference data to compute the likelihood}

\item{parallel}{either FALSE or number of nodes used in parallel processing}

\item{parallelOptions}{list with packages and objects to be exported to the cluster}

\item{pSnooker}{probability of Snooker update}

\item{burnin}{number of iterations treated as burn-in. These iterations are not recorded in the chain.}

\item{thin}{thinning parameter. Determines the interval in which values are recorded.}

\item{f}{scaling factor gamma}

\item{eps}{small number to avoid singularity}

\item{pGamma1}{probability determining the frequency with which the scaling is set to 1 (allows jumps between modes)}

\item{eps.mult}{random term (multiplicative error)}

\item{eps.add}{random term}

\item{zUpdateFrequency}{frequency with which the Z matrix is updated (= history of the chain)}

\item{numParticles}{number of particles}

\item{parallel}{either FALSE or number of nodes to be used in parallel processing}
}
\description{
Differential-Evolution MCMC zs with a particle filter to approximate the likelihood
}
\references{
ter  Braak C. J. F., and Vrugt J. A. (2008). Differential Evolution Markov Chain with snooker updater and fewer chains. Statistics and Computing http://dx.doi.org/10.1007/s11222-008-9104-9
}

