% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMCMC.R
\name{PMCMC}
\alias{PMCMC}
\title{Function to run the particle MCMC}
\usage{
PMCMC(iterations, model, likelihood, prior, f0, startValues, numParticles,
  scaling = NULL, observations, names = NULL, reportIntervall = 10,
  parallel = 2, parallelOptions = list(packages = NULL, objects = NULL))
}
\arguments{
\item{iterations}{number of iterations to run}

\item{model}{model to make predictions about the state}

\item{likelihood}{likelihood function (see details)}

\item{prior}{log prior density}

\item{f0}{function that generates a set of particles for start conditions (see details)}

\item{startValues}{vector containing the parameter start values. Alternatively
a function can be provided to sample the start values.}

\item{numParticles}{number of particles in the particle filter}

\item{scaling}{scaling factor for the proposal. If not provided 1/10 of the
prior with is used.}

\item{observations}{vactor containing the data (e.g. PAR) to run the model.}

\item{names}{parameter names (otional)}

\item{parallel}{number of nodes to be used}

\item{parallelOptions}{list with packages and objects to be exported to cluster}
}
\description{
Function to run the particle MCMC
}
\details{
The likelihood calculates the weigth for one set of particles. \cr\cr
The f0 function should accept the argument numParticles for the number of start particles.
}

