% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMCMC.R
\name{proposalFunction}
\alias{proposalFunction}
\title{Function to generate proposal}
\usage{
proposalFunction(parameter, scaling)
}
\arguments{
\item{parameter}{vector with current parameter values}

\item{scaling}{scaling factor for each parameter}
}
\value{
vactor with new proposal
}
\description{
Function to generate proposal
}

