% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{v3semC}
\alias{v3semC}
\title{C version of the VSEM model}
\usage{
v3semC(par, PAR)
}
\arguments{
\item{par}{parameter vector}

\item{PAR}{Photosynthetically active radiation (PAR) MJ /m2 /day}
}
\description{
C version of the VSEM model
}

